* dependency on balta library * database tests move to test folder from it * sbt - database tests are not run by default but by a devoted task
