+++ .gitignore
@@ -71,7 +71,7 @@ myTestCheckpoints
 **/main/resources/application.conf
 **/main/resources/application.yaml
 **/main/resources/application.yml
-/database/src/it/resources/database.properties
+/database/src/test/resources/database.properties
 
 *.config
 *.keytab
